Rexport SvelteComponent as SvelteComponent API
