+++ src/runtime/ambient.ts
@@ -1,19 +1,4 @@
 declare module '*.svelte' {
-	type Props = Record<string, any>;
-
-	export default class {
-		constructor(options: {
-			target: Element;
-			anchor?: Element;
-			props?: Props;
-			hydrate?: boolean;
-			intro?: boolean;
-		});
-
-		$set(props: Props): void;
-		$on<T = any>(event: string, callback: (event: CustomEvent<T>) => void): () => void;
-		$destroy(): void;
-
-		[accessor: string]: any;
-	}
+	type SvelteComponent = typeof import('./internal/ComponentApi').SvelteComponentApi
+	export default SvelteComponent
 }
